requirements.txt
rich
pypdf
pillow
openai
python-dotenv
anthropic
google-generativeai
groq
requests
pytest
